backend/books_exchange/requirements.txt
amqp==2.6.1
asgiref==3.3.1
billiard==3.6.3.0
celery==4.4.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.7.0rc2
Django==3.1.7
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
flower==0.9.7
gunicorn==20.0.4
humanize==3.2.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
kombu==4.6.11
MarkupSafe==1.1.1
oauthlib==3.1.0
prometheus-client==0.8.0
prompt-toolkit==3.0.16
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
tornado==6.1
uritemplate==3.0.1
urllib3==1.26.3
vine==1.3.0
wcwidth==0.2.5
